Initial Rune Rework
Changed implementation to use ScriptableObjects for the runes as well.

Things needed to be checked:
Will AnimatorControllers attached to ScriptableObjects control separate runes correctly, or will it be only one instance on all gameobjects.

Things to be added:
Animations for energy, Sprite sheets for different ranks of runes, OTher runes