I forgot to commit LOL
Finished moving over to scriptableobjects. There are a little too many nested classes I think for the pageData, however everything works.

the Buildsignalmanager is recieving signals, however the animations for the runes isn't done yet.

Oh wait lol, I forgot that I had to reformat the rune prefab a lot. The prefab is split into the base (displaying rune rank), the cover (displaying rune type), and the energy layers. The base and cover needed to be split as for hexagonal runes, the rotations didn't appear nice. Thus, the base will remain in the same orientation, while predrawn rotations of the cover will be drawn instead to ensure good quality images.

need to also add generating overworld projectiles for the output runes next, as well as attempting to figure out how to code triangle and hexagon runes.